Plugin UT: create structure for tests with etcd transport
